4688: Time/Date information is not updated after a Reset ocurrs.
The system shall take the GPS information from the Antenna connected when available, this is normal behavior but after a Reset the information is not taken until an Ignition Cycle occurs.

As a proposed solution and due to the lack of reproductivity, the missing onDestroy() method was implemented and the return statement of onStartCommand() was changed to START_REDELIVER_INTENT in order to keep the intent scheduled for re delivery in case it is killed by the system.